testing: promote shared test fixtures - uploadsOf, createSource, bitmap, video into src/testing/fixtures (internal, non-barrel)
