+++ src/effects/Grain/Grain.test.tsx
@@ -6,6 +6,7 @@ import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { Grain } from '@/effects/Grain/Grain';
 import type { GLStubHandle } from '@/testing';
 import { createGLStub } from '@/testing';
+import { uploadsOf } from '@/testing/fixtures';
 import type { FrameQueue } from '@/testing/frameQueue';
 import { createFrameQueue } from '@/testing/frameQueue';
 import type { Vec3 } from '@/types';
@@ -50,11 +51,6 @@ async function mount(element: ReactElement): Promise<void> {
     });
 }
 
-function uploadsOf(name: string): unknown[] {
-    const location = stub.gl.getUniformLocation({} as WebGLProgram, name);
-    return stub.uniformCalls.filter(call => call.location === location).map(call => call.value);
-}
-
 describe('Grain: real uniforms reach the GL stub and advance', () => {
     it('uploads the fixture colors and drives u_time forward across ticks', async () => {
         await mount(
@@ -74,17 +70,17 @@ describe('Grain: real uniforms reach the GL stub and advance', () => {
             act(() => { frames.tick(time) });
         }
 
-        const color = uploadsOf('u_color').map(value => Array.from(value as Float32Array));
+        const color = uploadsOf(stub, 'u_color').map(value => Array.from(value as Float32Array));
         expect(color.length).toBeGreaterThan(0);
         expect(color[0]).toEqual(Array.from(new Float32Array([0.09, 0.61, 0.37])));
 
-        const grainColor = uploadsOf('u_grainColor').map(value => Array.from(value as Float32Array));
+        const grainColor = uploadsOf(stub, 'u_grainColor').map(value => Array.from(value as Float32Array));
         expect(grainColor[0]).toEqual(Array.from(new Float32Array([0.88, 0.24, 0.66])));
 
-        const intensity = uploadsOf('u_intensity') as number[];
+        const intensity = uploadsOf(stub, 'u_intensity') as number[];
         expect(intensity).toContain(0.42);
 
-        const times = uploadsOf('u_time') as number[];
+        const times = uploadsOf(stub, 'u_time') as number[];
         expect(times.length).toBeGreaterThan(5);
         for (let i = 1; i < times.length; i++) {
             expect(times[i]).toBeGreaterThan(times[i - 1]);
